+++ test/tilt_redcarpettemplate_test.rb
@@ -9,13 +9,13 @@ begin
       assert_equal ['markdown', 'mkd', 'md'], extensions
     end
 
-    it "registered above BlueCloth" do
+    it "registered above Maruku" do
       %w[md mkd markdown].each do |ext|
         lazy = Tilt.lazy_map[ext]
-        blue_idx = lazy.index { |klass, file| klass == 'Tilt::BlueClothTemplate' }
+        maruku_idx = lazy.index { |klass, file| klass == 'Tilt::MarukuTemplate' }
         redc_idx = lazy.index { |klass, file| klass == 'Tilt::RedcarpetTemplate' }
-        assert redc_idx < blue_idx,
-          "#{redc_idx} should be lower than #{blue_idx}"
+        assert redc_idx < maruku_idx,
+          "#{redc_idx} should be lower than #{maruku_idx}"
       end
     end